requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciitree==0.3.3
astroid==3.2.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
babel==2.16.0
beautifulsoup4==4.12.3
black==24.4.2
bleach==6.1.0
build==1.2.1
cachetools==5.3.3
cads-api-client==1.3.0
Cartopy==0.23.0
cdsapi==0.7.2
certifi==2024.2.2
cffi==1.17.0
cfgv==3.4.0
cftime==1.6.4
charset-normalizer==3.3.2
check-shapes==1.1.1
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dask==2024.6.2
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.8
docker-pycreds==0.4.0
docutils==0.21.2
dropstackframe==0.1.0
ebird-api==3.0.6
einops==0.8.0
executing==2.0.1
fasteners==0.19
fastjsonschema==2.20.0
filelock==3.14.0
fonttools==4.52.4
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
gcsfs==2024.6.1
geodatasets==2024.7.0
geopandas==1.0.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.31.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
gpytorch==1.11
h11==0.14.0
h5netcdf==1.3.0
h5py==3.11.0
HiYaPyCo==0.6.0
httpcore==1.0.5
httpx==0.27.0
hydra-core==1.3.2
identify==2.5.36
idna==3.7
importlib_metadata==7.1.0
ipykernel==6.29.5
ipython==8.26.0
isoduration==20.11.0
isort==5.13.2
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaxtyping==0.2.29
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keyring==25.2.1
kiwisolver==1.4.5
lark==1.1.9
lightning==2.2.5
lightning-utilities==0.11.2
line_profiler==4.1.3
linear-operator==0.5.2
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.0.5
multiurl==0.3.1
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netCDF4==1.7.1
networkx==3.3
nh3==0.2.17
nodeenv==1.8.0
notebook_shim==0.2.4
numcodecs==0.12.1
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
pexpect==4.9.0
pillow==10.3.0
pkginfo==1.11.1
platformdirs==4.2.2
pooch==1.8.2
pre-commit==3.7.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
pylint==3.2.2
pyogrio==0.9.0
pyparsing==3.1.2
pyproj==3.6.1
pyproject_hooks==1.1.0
pyshp==2.3.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytorch-lightning==2.2.5
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.1.0
readme_renderer==43.0
referencing==0.35.1
requests==2.32.2
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
scikit-learn==1.5.0
scipy==1.13.1
Send2Trash==1.8.3
sentry-sdk==2.3.1
setproctitle==1.3.3
setuptools==69.5.1
shapely==2.0.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
sympy==1.12
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
tomlkit==0.12.5
toolz==0.12.1
torch==2.3.1
torchmetrics==1.4.0.post0
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
twine==5.1.0
typeguard==2.13.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
virtualenv==20.26.2
wandb==0.17.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.43.0
xarray==2024.6.0
yarl==1.9.4
zarr==2.18.2
zipp==3.19.2
